Add option (now default) for central differences
To replace the original mode of forward differences, which remains an
option.

Using central differences for the first derivative gives an additional
order of accuracy from discretisation, but doesn't necessarily help with
numerical stability (#3).

Test plan: There are new unit tests that test that both modes produce
output vectors of the correct length.